challenge-2025-06-18-with-crewai/pyproject.toml
[project]
name = "challenge-2025-06-18-with-crewai"
version = "0.1.0"
description = "Challenge with CrewAI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.121.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
